+++ src/core/retrieval/index.ts
@@ -9,7 +9,8 @@ export * from './interfaces/Retriever';
 export * from './interfaces/PromptComposer';
 
 // Implementations
-export * from './implementations/NoOpEmbedder';
+export * from './implementations/TransformersEmbedder';
+export * from './implementations/LocalEmbedder';
 export * from './implementations/InMemoryVectorStore';
 export * from './implementations/VectorStoreRetriever';
 export * from './implementations/DefaultPromptComposer';